+++ src/main/java/com/secucard/connect/product/general/model/AddressComponent.java
@@ -13,7 +13,6 @@
 package com.secucard.connect.product.general.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.io.Serializable;
 import java.util.List;