+++ src/main/java/com/secucard/connect/product/general/model/PublicMerchant.java
@@ -14,10 +14,10 @@ package com.secucard.connect.product.general.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.secucard.connect.product.common.model.SecuObject;
-
 import java.util.List;
 
 public class PublicMerchant extends SecuObject {
+
   @JsonProperty
   private String source;