+++ src/main/java/com/secucard/connect/product/loyalty/model/Sale.java
@@ -15,13 +15,13 @@ package com.secucard.connect.product.loyalty.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.secucard.connect.product.common.model.SecuObject;
 import com.secucard.connect.product.general.model.Store;
-
 import java.io.Serializable;
 import java.util.Currency;
 import java.util.Date;
 import java.util.List;
 
 public class Sale extends SecuObject {
+
   @JsonProperty
   private int amount;
 
@@ -165,6 +165,7 @@ public class Sale extends SecuObject {
 
 
   public static class Bonus implements Serializable {
+
     private int amount;
 
     /**