+++ src/main/java/com/secucard/connect/product/payment/model/RedirectUrl.java
@@ -14,34 +14,45 @@ public class RedirectUrl {
   @JsonProperty("iframe_url")
   public String iframeUrl;  // The url for redirect the customer to the payment checkout page
 
+  @JsonProperty("url_push")
+  public String urlPush;  // The url for receiving push notifications
+
   public String getUrlSuccess() {
     return urlSuccess;
   }
 
-  public void setUrlSuccess(String url_success) {
-    this.urlSuccess = url_success;
+  public void setUrlSuccess(String urlSuccess) {
+    this.urlSuccess = urlSuccess;
   }
 
   public String getUrlFailure() {
     return urlFailure;
   }
 
-  public void setUrlFailure(String url_failure) {
-    this.urlFailure = url_failure;
+  public void setUrlFailure(String urlFailure) {
+    this.urlFailure = urlFailure;
   }
 
   public String getUrlIframe() {
     return iframeUrl;
   }
 
-  public void setUrlIframe(String iframe_url) {
-    this.iframeUrl = iframe_url;
+  public void setUrlIframe(String iframeUrl) {
+    this.iframeUrl = iframeUrl;
+  }
+
+  public String getUrlPush() {
+    return urlPush;
+  }
+
+  public void setUrlPush(String urlPush) {
+    this.urlPush = urlPush;
   }
 
   @Override
   public String toString() {
-    return "RedirectUrl{" + "urlSuccess='" + urlSuccess + '\'' + "urlFailure='" + urlFailure + '\''
-        + "iframeUrl='" + iframeUrl + '\'' + '}';
+    return "RedirectUrl{" + "urlSuccess='" + urlSuccess + '\'' + ", urlFailure='" + urlFailure + '\'' + ", iframeUrl='" + iframeUrl + '\''
+        + ", url_push='" + urlPush + '\'' + '}';
   }
 
 }