+++ src/main/java/com/secucard/connect/product/payment/model/Transaction.java
@@ -67,6 +67,8 @@ public abstract class Transaction extends SecuObject {
   @JsonProperty("transaction_status")
   protected String transactionStatus;
 
+  protected boolean demo;
+
   public Customer getCustomer() {
     return customer;
   }
@@ -222,6 +224,14 @@ public abstract class Transaction extends SecuObject {
   }
   public void setIframeUrl(String iframeUrl) {
     this.iframeUrl = iframeUrl;
+  }
+
+  public boolean getDemo() {
+    return demo;
+  }
+
+  public void setDemo(boolean demo) {
+    this.demo = demo;
   }
 
 }