+++ src/main/java/com/secucard/connect/product/smart/model/CashierDisplay.java
@@ -59,10 +59,6 @@ public class CashierDisplay {
 
   @Override
   public String toString() {
-    return "CashierDisplay{" +
-        "deviceId='" + deviceId + '\'' +
-        ", action='" + action + '\'' +
-        ", value='" + value + '\'' +
-        '}';
+    return "CashierDisplay{" + "deviceId='" + deviceId + '\'' + ", action='" + action + '\'' + ", value='" + value + '\'' + '}';
   }
 }